renders datasets overview page successfully
