+++ app/blog/utils/mdx.ts
@@ -61,8 +61,8 @@ export function resolveConfigFunctions(
 function parseAttributes(obj) {
   const convert = (obj) => {
     return Object.keys(obj).reduce((acc, key) => {
-      if (typeof obj[key] === 'object' && obj[key] !== null) {
-        acc[key] = Array.isArray(obj[key]) ? obj[key].map(convert) : convert(obj[key]);
+      if (typeof obj[key] === 'object' && obj[key] !== null) {     
+        acc[key] = convert(obj[key]);
       } else if (typeof obj[key] === 'string') {
 
         if (obj[key].includes('::markdown')) {
@@ -152,41 +152,4 @@ export function getStories() {
 
 export function getBlogPosts() {
   return getMDXData(path.join(process.cwd(), 'app', 'blog', 'datasets'))
-}
-
-export function formatDate(date: string, includeRelative = false) {
-  let currentDate = new Date()
-  if (!date || typeof date !== 'string') return;
-  if (!date.includes('T')) {
-    date = `${date}T00:00:00`
-  }
-  let targetDate = new Date(date)
-
-  let yearsAgo = currentDate.getFullYear() - targetDate.getFullYear()
-  let monthsAgo = currentDate.getMonth() - targetDate.getMonth()
-  let daysAgo = currentDate.getDate() - targetDate.getDate()
-
-  let formattedDate = ''
-
-  if (yearsAgo > 0) {
-    formattedDate = `${yearsAgo}y ago`
-  } else if (monthsAgo > 0) {
-    formattedDate = `${monthsAgo}mo ago`
-  } else if (daysAgo > 0) {
-    formattedDate = `${daysAgo}d ago`
-  } else {
-    formattedDate = 'Today'
-  }
-
-  let fullDate = targetDate.toLocaleString('en-us', {
-    month: 'long',
-    day: 'numeric',
-    year: 'numeric',
-  })
-
-  if (!includeRelative) {
-    return fullDate
-  }
-
-  return `${fullDate} (${formattedDate})`
 }
