+++ app/datasets/[slug]/page.tsx
@@ -1,7 +1,8 @@
 import { notFound } from 'next/navigation'
 import { CustomMDX } from 'app/components/mdx'
-import { formatDate, getBlogPosts, getDatasets } from 'app/blog/utils'
+import { getDatasets } from 'app/blog/utils/mdx';
 import { baseUrl } from 'app/sitemap'
+import {PageHero} from 'app/lib'
 
 function generateStaticParams() {
   let posts = getDatasets()
@@ -13,7 +14,6 @@ function generateStaticParams() {
 
 export function generateMetadata({ params }) {
   let post = getDatasets().find((post) => post.slug === params.slug)
-  // console.log(post)
   
   if (!post) {
     return
@@ -53,42 +53,20 @@ export function generateMetadata({ params }) {
 
 export default function Blog({ params }) {
   let post = getDatasets().find((post) => post.slug === params.slug)
-  console.log(post);
+
   if (!post) {
     notFound()
   }
 
   return (
     <section>
-      <script
-        type="application/ld+json"
-        suppressHydrationWarning
-        dangerouslySetInnerHTML={{
-          __html: JSON.stringify({
-            '@context': 'https://schema.org',
-            '@type': 'BlogPosting',
-            headline: post.metadata.name,
-            datePublished: post.metadata.pubDate,
-            dateModified: post.metadata.pubDate,
-            description: post.metadata.description,
-            url: `${baseUrl}/blog/${post.slug}`,
-            author: {
-              '@type': 'Person',
-              name: 'My Portfolio',
-            },
-          }),
-        }}
-      />
-      <h1 className="title font-semibold text-2xl tracking-tighter">
-        {post.metadata.title}
-      </h1>
-      <div className="flex justify-between items-center mt-2 mb-8 text-sm">
-        <p className="text-sm text-neutral-600 dark:text-neutral-400">
-          {formatDate(post.metadata.publishedAt)}
-        </p>
-      </div>
-      <h3> hihi</h3>
       <article className="prose">
+      <PageHero
+        title={post.metadata.name}
+        description={post.metadata.description}
+        coverSrc={post.metadata.media?.src} 
+        coverAlt={post.metadata.media?.alt}
+      />
         <CustomMDX source={post.content} />
       </article>
     </section>
