+++ app/layout.tsx
@@ -1,14 +1,11 @@
 import './global.css'
-import type { Metadata } from 'next'
-import { GeistSans } from 'geist/font/sans'
-import { GeistMono } from 'geist/font/mono'
-import { Navbar } from './components/nav'
-import { Analytics } from '@vercel/analytics/react'
-import { SpeedInsights } from '@vercel/speed-insights/next'
-import Footer from './components/footer'
-import { baseUrl } from './sitemap'
-import { DataProvider } from './store/provider'
-import { getDatasets } from './blog/utils'
+import type { Metadata } from 'next';
+import { Navbar } from './components/nav';
+import { Analytics } from '@vercel/analytics/react';
+import { SpeedInsights } from '@vercel/speed-insights/next';
+import Footer from './components/footer';
+import { baseUrl } from './sitemap';
+import { CoreVedaProvidersWrapper } from './store/providers';
 
 export const metadata: Metadata = {
   metadataBase: new URL(baseUrl),
@@ -48,18 +45,13 @@ export default function RootLayout({
   return (
     <html
       lang="en"
-      className={cx(
-        'text-black bg-white dark:text-white dark:bg-black',
-        GeistSans.variable,
-        GeistMono.variable
-      )}
     >
-      <body className="antialiased max-w-xl mx-4 mt-8 lg:mx-auto">
-        <main className="flex-auto min-w-0 mt-6 flex flex-col px-2 md:px-0">
+      <body>
+        <main>
           <Navbar />
-          <DataProvider initialDatasets={getDatasets()}>
+          <CoreVedaProvidersWrapper>
             {children}
-          </DataProvider>
+          </CoreVedaProvidersWrapper>
           <Footer />
           <Analytics />
           <SpeedInsights />
