+++ app/lib/index.ts
@@ -1,0 +1,6 @@
+'use client';
+import { PageHero } from '@developmentseed/veda-ui';
+
+export {
+  PageHero
+}
